Use str_to_byte_list() for both i2c_base.py and uart_base.py
